Send terminal reset string directly rather than calling reset(1)
reset(1) doesn't actually send a reset string, but uses ioctl() to talk
to reset the tty. 'tput reset' sends a reset string tailored to the
terminal in use. Rather than depend on tput, we make the
probably-correct assumption that the vt220 "full reset" code will be
accepted by most terminals.