Total overhaul of closed caption data file generation
The data files are no longer included in the source; instead they're
generated as part of the build process using DRL's soundhq.lua
configuration. As a side effect of this, a number of sound effects have
changed, and support for some new ones has been added.

This also begins laying the groundwork for actual sound playback in the
web terminal.